// src/custom_types/enums.ts
export enum ProjectStatuses {
    TO_DO = "TO_DO",
    RESOLVED = "RESOLVED",
    CLOSED = "CLOSED",
    IN_PROGRESS = "IN_PROGRESS"
}

export enum TaskTypes {
    STORY = "STORY",
    TASK = "TASK",
    BUGFIX = "BUGFIX"
}

export enum TaskStatuses {
    TO_DO = "TO_DO",
    RESOLVED = "RESOLVED",
    CLOSED = "CLOSED",
    IN_PROGRESS = "IN_PROGRESS"
}